Move hoversupport module to gaphor.diagram
So modeling languages do not depend on gaphor.ui.
