+++ gaphor/UML/classes/associationpropertypages.py
@@ -1,8 +1,8 @@
 from gaphor import UML
 from gaphor.core.format import format, parse
+from gaphor.diagram.hoversupport import widget_add_hover_support
 from gaphor.diagram.propertypages import PropertyPageBase, PropertyPages
 from gaphor.transaction import transactional
-from gaphor.ui.hoversupport import widget_add_hover_support
 from gaphor.UML.classes.association import AssociationItem
 from gaphor.UML.classes.classespropertypages import new_builder
 from gaphor.UML.profiles.stereotypepropertypages import stereotype_model
