+++ gaphor/UML/classes/classespropertypages.py
@@ -5,6 +5,7 @@ from gi.repository import Gdk, Gtk
 from gaphor import UML
 from gaphor.core import gettext, transactional
 from gaphor.core.format import format, parse
+from gaphor.diagram.hoversupport import widget_add_hover_support
 from gaphor.diagram.propertypages import (
     EditableTreeModel,
     PropertyPageBase,
@@ -13,7 +14,6 @@ from gaphor.diagram.propertypages import (
     on_bool_cell_edited,
     on_text_cell_edited,
 )
-from gaphor.ui.hoversupport import widget_add_hover_support
 from gaphor.UML.classes.datatype import DataTypeItem
 from gaphor.UML.classes.interface import Folded, InterfaceItem
 from gaphor.UML.classes.klass import ClassItem
