+++ gaphor/diagram/hoversupport.py
